# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Decision Brain API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /decisions:
    get:
      operationId: listDecisions
      tags: [decisions]
      summary: List decisions with filters
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: platform
          in: query
          schema:
            type: string
        - name: stakes
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, reversed, pending]
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of decisions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionListResponse"
    post:
      operationId: createDecision
      tags: [decisions]
      summary: Manually log a decision
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDecisionBody"
      responses:
        "201":
          description: Decision created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Decision"
  /decisions/{id}:
    get:
      operationId: getDecision
      tags: [decisions]
      summary: Get a decision by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Decision detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionDetail"
    patch:
      operationId: updateDecision
      tags: [decisions]
      summary: Update a decision
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDecisionBody"
      responses:
        "200":
          description: Updated decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Decision"
    delete:
      operationId: deleteDecision
      tags: [decisions]
      summary: Delete a decision
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /decisions/{id}/similar:
    get:
      operationId: getSimilarDecisions
      tags: [decisions]
      summary: Get semantically similar decisions
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Similar decisions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Decision"
  /outcomes:
    post:
      operationId: createOutcome
      tags: [outcomes]
      summary: Log an outcome for a decision
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOutcomeBody"
      responses:
        "201":
          description: Outcome created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Outcome"
  /outcomes/pending:
    get:
      operationId: getPendingOutcomes
      tags: [outcomes]
      summary: Get decisions awaiting outcome review
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Pending outcomes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PendingOutcome"
  /outcomes/{id}:
    patch:
      operationId: updateOutcome
      tags: [outcomes]
      summary: Update an outcome
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOutcomeBody"
      responses:
        "200":
          description: Updated outcome
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Outcome"
...
  /integrations:
    get:
      operationId: listIntegrations
      tags: [integrations]
      summary: List platform integrations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of integrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Integration"
  /integrations/{platform}/connect:
    post:
      operationId: connectIntegration
      tags: [integrations]
      summary: Connect a platform integration
      security:
        - bearerAuth: []
      parameters:
        - name: platform
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Connection initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Integration"
  /integrations/{platform}/disconnect:
    delete:
      operationId: disconnectIntegration
      tags: [integrations]
      summary: Disconnect a platform integration
      security:
        - bearerAuth: []
      parameters:
        - name: platform
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Disconnected
...
  /analytics/outcomes:
    get:
      operationId: getOutcomeAnalytics
      tags: [analytics]
      summary: Outcome performance analytics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Outcome analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutcomeAnalytics"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Decision:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        sourcePlatform:
          type: string
          nullable: true
          enum: [gmail, zoom, slack, meet, teams, notion, manual, null]
        stakes:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [active, reversed, pending]
        tags:
          type: array
          items:
            type: string
        decidedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        outcomeScore:
          type: integer
          nullable: true
        alertCount:
          type: integer
      required:
        - id
        - userId
        - title
        - stakes
        - status
        - tags
        - createdAt
        - alertCount
    DecisionDetail:
      allOf:
        - $ref: "#/components/schemas/Decision"
        - type: object
          properties:
            rawContext:
              type: string
              nullable: true
            outcomes:
              type: array
              items:
                $ref: "#/components/schemas/Outcome"
            alerts:
              type: array
              items:
                $ref: "#/components/schemas/Alert"
          required:
            - outcomes
            - alerts
    DecisionListResponse:
      type: object
      properties:
        decisions:
          type: array
          items:
            $ref: "#/components/schemas/Decision"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - decisions
        - total
        - page
        - limit
        - totalPages
    CreateDecisionBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        stakes:
          type: string
          enum: [low, medium, high, critical]
          default: medium
        tags:
          type: array
          items:
            type: string
        decidedAt:
          type: string
          format: date-time
        sourcePlatform:
          type: string
      required:
        - title
    UpdateDecisionBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        stakes:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [active, reversed, pending]
        tags:
          type: array
          items:
            type: string
    Outcome:
      type: object
      properties:
        id:
          type: string
        decisionId:
          type: string
        userId:
          type: string
        score:
          type: integer
          minimum: 0
          maximum: 100
        notes:
          type: string
          nullable: true
        checkInterval:
          type: string
          enum: ["30d", "90d", "180d", "1y"]
        trackedAt:
          type: string
          format: date-time
      required:
        - id
        - decisionId
        - userId
        - score
        - checkInterval
        - trackedAt
    PendingOutcome:
      type: object
      properties:
        decision:
          $ref: "#/components/schemas/Decision"
        dueAt:
          type: string
          format: date-time
        checkInterval:
          type: string
      required:
        - decision
        - dueAt
        - checkInterval
    CreateOutcomeBody:
      type: object
      properties:
        decisionId:
          type: string
        score:
          type: integer
          minimum: 0
          maximum: 100
        notes:
          type: string
        checkInterval:
          type: string
          enum: ["30d", "90d", "180d", "1y"]
          default: "90d"
      required:
        - decisionId
        - score
    UpdateOutcomeBody:
      type: object
      properties:
        score:
          type: integer
          minimum: 0
          maximum: 100
        notes:
          type: string
    Alert:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        decisionId:
          type: string
          nullable: true
        alertType:
          type: string
          enum: [pattern_match, blind_spot, repeat_mistake, similarity]
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [low, medium, high]
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - alertType
        - title
        - description
        - severity
        - isRead
        - createdAt
    Integration:
      type: object
      properties:
        id:
          type: string
        platform:
          type: string
          enum: [gmail, zoom, slack, meet, teams, outlook, notion, docusign]
        status:
          type: string
          enum: [connected, disconnected, error]
        lastSyncedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - platform
        - status
        - createdAt
...
    OutcomeAnalytics:
      type: object
      properties:
        winRate:
          type: number
        avgScore:
          type: number
        scoreByStakes:
          type: object
          additionalProperties:
            type: number
        scoreByPlatform:
          type: object
          additionalProperties:
            type: number
        scoreTrend:
          type: array
          items:
            $ref: "#/components/schemas/TrendPoint"
        topDecisions:
          type: array
          items:
            $ref: "#/components/schemas/Decision"
      required:
        - winRate
        - avgScore
        - scoreByStakes
        - scoreByPlatform
        - scoreTrend
        - topDecisions
...
    TrendPoint:
      type: object
      properties:
        date:
          type: string
        value:
          type: number
      required:
        - date
        - value
...